+++ eras/conway/impl/testlib/Test/Cardano/Ledger/Conway/Imp/LedgerSpec.hs
@@ -6,6 +6,7 @@
 {-# LANGUAGE ScopedTypeVariables #-}
 {-# LANGUAGE TypeApplications #-}
 {-# LANGUAGE TypeFamilies #-}
+{-# LANGUAGE TypeOperators #-}
 
 module Test.Cardano.Ledger.Conway.Imp.LedgerSpec (spec) where
 
@@ -22,13 +23,16 @@ import Cardano.Ledger.Conway.Rules (
   ConwayGovPredFailure (UnelectedCommitteeVoters),
   ConwayLedgerPredFailure (..),
   ConwayUtxoPredFailure (BadInputsUTxO),
+  PredicateFailure,
  )
 import Cardano.Ledger.Credential (Credential (..))
 import Cardano.Ledger.DRep
 import Cardano.Ledger.Plutus (SLanguage (..), hashPlutusScript)
-import Cardano.Ledger.Shelley.API.Mempool (ApplyTxError (..), applyTx, mkMempoolEnv)
+import Cardano.Ledger.Shelley.API.Mempool (applyTx, mkMempoolEnv)
 import Cardano.Ledger.Shelley.LedgerState
 import Control.Monad.Reader (asks)
+import Data.List.NonEmpty (NonEmpty)
+import qualified Data.List.NonEmpty as NonEmpty
 import qualified Data.Map.Strict as Map
 import qualified Data.Set as Set
 import qualified Data.Text as T
@@ -212,6 +216,11 @@ spec = do
 
   describe "Mempool" $ do
     let
+      submitFailingMempoolTx ::
+        String ->
+        Tx TopTx era ->
+        NonEmpty (PredicateFailure (EraRule "LEDGER" era)) ->
+        ImpM (LedgerSpec era) (Tx TopTx era)
       submitFailingMempoolTx cause tx expectedFailures = do
         globals <- use impGlobalsL
         nes <- use impNESL
@@ -223,10 +232,16 @@ spec = do
         logToExpr txFixed
         case applyTx globals mempoolEnv ls txFixed of
           Left err -> do
-            err `shouldBe` ApplyTxError @era expectedFailures
+            err `shouldBe` inject expectedFailures
           Right _ ->
             assertFailure $ "Expected failure due to " <> cause <> ": " <> show txFixed
         pure txFixed
+
+      submitFailingMempoolTx_ ::
+        String ->
+        Tx TopTx era ->
+        NonEmpty (PredicateFailure (EraRule "LEDGER" era)) ->
+        ImpM (LedgerSpec era) ()
       submitFailingMempoolTx_ c t = void . submitFailingMempoolTx c t
 
     it "Duplicate transactions" $ do
@@ -248,7 +263,7 @@ spec = do
       impAnn "Identical transaction" $ do
         withNoFixup $
           submitFailingMempoolTx_ "duplicate transaction" txFinal $
-            pure . injectFailure . ConwayMempoolFailure $
+            NonEmpty.singleton . injectFailure . ConwayMempoolFailure $
               "All inputs are spent. Transaction has probably already been included"
 
       impAnn "Overlapping transaction" $ do